# api/http/swagger.yaml
basePath: /v1
definitions:
...
  model.UserFollow:
    properties:
      fans_num:
        description: 粉丝数
        type: integer
      follow_num:
        description: 关注数
        type: integer
      is_fans:
        description: 是否是粉丝 1:是 0:否
        type: integer
      is_follow:
        description: 是否关注 1:是 0:否
        type: integer
    type: object
  model.UserInfo:
    properties:
      avatar:
        type: string
      id:
        example: 1
        type: integer
      sex:
        type: integer
      user_follow:
        $ref: '#/definitions/model.UserFollow'
      username:
        example: 张三
        type: string
    type: object
  user.LoginCredentials:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.PhoneLoginCredentials:
    properties:
      phone:
        example: 13010002000
        type: integer
      verify_code:
        example: 120110
        type: integer
    required:
    - phone
    - verify_code
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: eagle demo
  title: eagle docs api
  version: "1.0"
paths:
...
  /login:
    post:
      description: 仅限邮箱登录
      parameters:
      - description: 请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.LoginCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/model.UserInfo'
      summary: 用户登录接口
      tags:
      - 用户
...
  /users/login:
    post:
      description: 仅限手机登录
      parameters:
      - description: phone
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.PhoneLoginCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/model.UserInfo'
      summary: 用户登录接口
      tags:
      - 用户
...
swagger: "2.0"